Version 1.2
- Added Set Microphone Volume (Loop to 70%) option
Note:
DeviceNumber will vary from system to system. Please run the soundcard analysis script to find out what the DeviceNumber of your microphone is. This script can be found in the AHK help here: https://autohotkey.com/docs/commands/SoundSet.htm#Ex . The soundcard analysis script is not very descriptive, so a good way to glean the information you require is to set the volume of the mic you are interested to something weird, like 57, from within the Windows recording devices Properties interface. Then, run the soundcard analysis script and look for the entry with the Setting column matching the volume you set. Note the number in the Mixer column. This number = DeviceNumber in this script. Substitute your DeviceNumber in this script.